IGNITE-16566 Move IgniteNameUtils and StringUtils to api module
