+++ modules/api/src/main/java/org/apache/ignite/lang/ColumnNotFoundException.java
@@ -17,10 +17,9 @@
 
 package org.apache.ignite.lang;
 
-import static org.apache.ignite.internal.util.IgniteNameUtils.canonicalName;
-import static org.apache.ignite.internal.util.IgniteNameUtils.quote;
 import static org.apache.ignite.lang.ErrorGroups.Table.COLUMN_NOT_FOUND_ERR;
-import static org.apache.ignite.lang.IgniteStringFormatter.format;
+import static org.apache.ignite.lang.util.IgniteNameUtils.canonicalName;
+import static org.apache.ignite.lang.util.IgniteNameUtils.quote;
 
 import java.util.UUID;
 
@@ -34,7 +33,7 @@ public class ColumnNotFoundException extends IgniteException {
      * @param columnName Column name.
      */
     public ColumnNotFoundException(String columnName) {
-        super(COLUMN_NOT_FOUND_ERR, format("Column does not exist [name={}]", quote(columnName)));
+        super(COLUMN_NOT_FOUND_ERR, "Column does not exist [name=" + quote(columnName) + ']');
     }
 
     /**
@@ -45,8 +44,9 @@ public class ColumnNotFoundException extends IgniteException {
      * @param tableName Table name.
      */
     public ColumnNotFoundException(String schemaName, String tableName, String columnName) {
-        super(COLUMN_NOT_FOUND_ERR, format("Column does not exist [tableName={}, columnName={}]",
-                canonicalName(schemaName, tableName), quote(columnName)));
+        super(
+                COLUMN_NOT_FOUND_ERR,
+                "Column does not exist [tableName=" + canonicalName(schemaName, tableName) + ", columnName=" + quote(columnName) + ']');
     }
 
     /**
