+++ modules/api/src/main/java/org/apache/ignite/lang/IndexAlreadyExistsException.java
@@ -17,9 +17,8 @@
 
 package org.apache.ignite.lang;
 
-import static org.apache.ignite.internal.util.IgniteNameUtils.canonicalName;
 import static org.apache.ignite.lang.ErrorGroups.Index.INDEX_ALREADY_EXISTS_ERR;
-import static org.apache.ignite.lang.IgniteStringFormatter.format;
+import static org.apache.ignite.lang.util.IgniteNameUtils.canonicalName;
 
 import java.util.UUID;
 
@@ -34,7 +33,7 @@ public class IndexAlreadyExistsException extends IgniteException {
      * @param indexName Index name.
      */
     public IndexAlreadyExistsException(String schemaName, String indexName) {
-        super(INDEX_ALREADY_EXISTS_ERR, format("Index already exists [name={}]", canonicalName(schemaName, indexName)));
+        super(INDEX_ALREADY_EXISTS_ERR, "Index already exists [name=" + canonicalName(schemaName, indexName) + ']');
     }
 
     /**
