+++ modules/api/src/main/java/org/apache/ignite/table/mapper/OneColumnMapperImpl.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.table.mapper;
 
-import org.apache.ignite.internal.util.IgniteNameUtils;
+import org.apache.ignite.lang.util.IgniteNameUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
