+++ modules/client/src/main/java/org/apache/ignite/internal/client/table/ClientTuple.java
@@ -18,7 +18,7 @@
 package org.apache.ignite.internal.client.table;
 
 import org.apache.ignite.internal.binarytuple.BinaryTupleReader;
-import org.apache.ignite.internal.util.IgniteNameUtils;
+import org.apache.ignite.lang.util.IgniteNameUtils;
 import org.apache.ignite.sql.ColumnType;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
