+++ modules/core/src/test/java/org/apache/ignite/lang/IgniteExceptionMapperUtilTest.java
