+++ modules/core/src/test/java/org/apache/ignite/lang/TestIgniteExceptionMappersProvider.java
